+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `WalletConnection` type
 - Added `transferWithSigner` function to enable transfer with l2signer
 - Added `batchNftTransferWithSigner` function to enable batch transfer with l2signer
+- Added `prepareWithdrawalWorkflowWithSigner` function to enable prepare withdrawal with l2signer
 
 ### Deprecated
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `cancelOrder`, use `cancelOrderWithSigner` instead
 - `transfer`, use `transferWithSigner` instead
 - `batchNftTransfer`, use `batchNftTransferWithSigner` instead
+- `prepareWithdrawalWorkflow`, use `prepareWithdrawalWorkflowWithSigner` instead
 
 ## [0.5.0] - 2022-07-12